Move all classes to separate files
