+++ src/Collections/BlockSparseArray.cs
@@ -140,91 +140,3 @@ public class BlockSparseArray<T>
         Array.Clear(_blocks, 0, _blocks.Length);
     }
 }
-
-class SparseBlock<T>
-{
-    private readonly T[] _block;
-
-    public SparseBlock(int startIndex, int length)
-    {
-        if (startIndex < 0) throw new ArgumentException("Start Index must be >= 0", nameof(startIndex));
-        if (length <= 0) throw new ArgumentException("Length must be >= 1", nameof(length));
-        StartIndex = startIndex;
-        Length = length;
-        _block = new T[length];
-    }
-
-    public int StartIndex { get; private set; }
-
-    public int Length { get; private set; }
-
-    public T this[int index]
-    {
-        get => _block[index - StartIndex];
-        set => _block[index - StartIndex] = value;
-    }
-}
-
-class BlockSparseArrayEnumerator<T>
-    : IEnumerator<T>
-{
-    public BlockSparseArrayEnumerator(IEnumerator blocks, int length, int blockSize)
-    {
-        Blocks = blocks;
-        Length = length;
-        Index = -1;
-        BlockSize = blockSize;
-    }
-
-    private IEnumerator Blocks { get; set; }
-
-    private int Index { get; set; }
-
-    private int Length { get; set; }
-
-    private int BlockSize { get; set; }
-
-    public void Dispose()
-    {
-        // No dispose actions needed
-    }
-
-    public bool MoveNext()
-    {
-        if (Index == -1)
-        {
-            Blocks.MoveNext();
-        }
-        Index++;
-        if (Index > 0 && Index % BlockSize == 0)
-        {
-            // Need to move to next block
-            if (!Blocks.MoveNext()) return false;
-        }
-        return Index < Length;
-    }
-
-    public void Reset()
-    {
-        Index = -1;
-        Blocks.Reset();
-    }
-
-    public T Current
-    {
-        get
-        {
-            if (Index == -1) throw new InvalidOperationException("Currently before the start of the enumerator, please call MoveNext() before accessing this property");
-            if (Index >= Length) throw new InvalidOperationException("Past the end of the enumerator");
-
-            // If no current block return default value
-            var currentBlock = (SparseBlock<T>) Blocks.Current;
-            if (currentBlock == null) return default(T);
-
-            // Otherwise return the value within the block
-            return currentBlock[Index];
-        }
-    }
-
-    object IEnumerator.Current => Current;
-}
